requirements.txt
matplotlib
requests
numpy
fastapi[standard]
h3
python-dotenv
rasterio
uvicorn